pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel>=0.41"]
build-backend = "setuptools.build_meta"

[project]
name = "wjp-analyser"
version = "0.1.0"
description = "DXF validator/analyzer/visualizer/G-code for waterjet"
requires-python = ">=3.10"
dependencies = [
  "ezdxf>=1.3.0",
  "flask>=3.0.3",
  "streamlit>=1.34.0",
  "matplotlib>=3.9.0",
  "numpy>=1.26.4",
  "opencv-python>=4.10.0",
  "pillow>=10.4.0",
  "pydantic>=2.7.0",
  "rich>=13.7.1",
  "scikit-image>=0.24.0",
  "shapely>=2.0.2",
  "pandas>=2.0.0",
  "plotly>=5.24.0",
  "click>=8.1.0",
  "openai>=1.0.0",
  "pyyaml>=6.0",
  "requests>=2.25.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "rectpack>=0.2.2",
  "svgpathtools>=1.6.1",
  "python-magic>=0.4.27",
  "sqlalchemy>=2.0.0",
  "redis>=5.0.0",
  "rq>=1.15.0",
  "pyjwt>=2.4.0",
  "reportlab>=4.1.0",
  "seaborn>=0.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"wjp_analyser.web" = [
  "templates/*.html",
  "static/*",
]

[project.scripts]
wjdx = "cli.main:main"
wjdx-web = "wjp_analyser.cli.web:main"
wjdx-layerify = "wjp_analyser.cli.layerify:main"
wjdx-gcode = "wjp_analyser.gcode.cli:main"
